First implementation of central_tuplespace, still has running problem.
contains following commits.
{
add managed_central_tuplespace example codes, not tested.

fix compilation errors and runs.

use vector instead of deque.

create storage/tuple.hpp.

overload == operator in small_big_objects.hpp

use hpx::util::any instead of hpx::util::hold_any.

avoid DEBUG name conflict.

fix tuplespace annoying ambiguous warnings.

switch to simple_component and some fix.

use actions and agas register.

fix vs2012 errors.
}

+++ hpx/util/one_size_heap_list.hpp
@@ -60,7 +60,7 @@ namespace hpx { namespace util
             , max_alloc_count_(0L)
 #endif
         {
-            BOOST_ASSERT(sizeof(typename heap_type::storage_type) == heap_size);
+            BOOST_ASSERT(sizeof(typename heap_type::storage_type) == uint64_t(heap_size));
         }
 
         explicit one_size_heap_list(std::string const& class_name)
@@ -72,7 +72,7 @@ namespace hpx { namespace util
             , max_alloc_count_(0L)
 #endif
         {
-            BOOST_ASSERT(sizeof(typename heap_type::storage_type) == heap_size);
+            BOOST_ASSERT(sizeof(typename heap_type::storage_type) == uint64_t(heap_size));
         }
 
         ~one_size_heap_list()
